Correccion estructura datos y consultas

+++ auth0rize.auth.domain/Application/IApplicationRepository.cs
@@ -4,6 +4,6 @@ namespace auth0rize.auth.domain.Application
 {
     public interface IApplicationRepository
     {
-        Task<List<ApplicationGet>> get(int userId);
+        Task<List<ApplicationGet>> get(long userId);
     }
 }
